+++ frontend/src/@types/common.d.ts
@@ -4,9 +4,4 @@ type CommonType = {
   updated_at: string;
 };
 
-export type LoginForm = {
-  email: string;
-  password: string;
-}
-
 export default CommonType;
