--- frontend/src/components/LoginForm/LoginForm.type.ts
@@ -1,12 +1,0 @@
-import { FormInstance } from 'antd';
-import { LoginForm } from '@types';
-
-interface Props {
-  initialState: LoginForm;
-  onFinish: (values: any) => void;
-  onFinishFailed: (values: any) => void;
-  form: FormInstance<any>;
-  isLoading: boolean;
-}
-
-export default Props;
