+++ frontend/src/components/LoginFormulary/index.ts
@@ -1,0 +1,1 @@
+export { default } from './LoginFormulary';
